Fixed narrowing message of PolygonDifferences and also changed to direct initialization instead of uniform because of compile error on gcc

+++ include/lane_detect_processor.h
@@ -34,7 +34,7 @@ struct EvaluatedLine {
 
 struct PolygonDifferences {
 	Polygon polygon;
-	float differencefromaverage;
+	int differencefromaverage;
 };
 
 bool CheckAngle( const cv::Point center,
